move handleSetupContentfulClient to its own file
